:building_construction: refactor: Heatmapタスクページを改善
Heatmapタスクページをリファクタリングし、再利用可能なコンポーネントへ変更。APIクエリ処理やレンダリング構造を整理し、コードの読みやすさと保守性を向上。
